dnsmon-build/requirements.txt
dnspython
elasticsearch==8.9.0


